Proyectos Extra/Programacion/Omni_ERP/pyproject.toml
[build-system]
requires = ["hatchling>=1.25.0"]
build-backend = "hatchling.build"

[project]
name = "dario-app"
version = "0.1.0"
description = "Proyecto base en Python con estructura src y tests"
readme = "README.md"
requires-python = ">=3.9"
authors = [
  { name = "Dario" }
]
license = { text = "Proprietary" }
keywords = ["erp", "fastapi", "sqlalchemy", "web"]
dependencies = [
  "fastapi>=0.115.0",
  "uvicorn[standard]>=0.32.0",
  "sqlalchemy>=2.0.0",
  "alembic>=1.13.0",
  "pydantic[email]>=2.0.0",
  "pydantic-settings>=2.0.0",
  "python-multipart>=0.0.9",
  "jinja2>=3.1.0",
  "aiosqlite>=0.20.0",
  "passlib[bcrypt]>=1.7.4",
  "python-jose[cryptography]>=3.3.0"
]

[project.scripts]
dario = "dario_app.cli:main"
dario-server = "dario_app.server:run_server"

[project.optional-dependencies]
dev = [
  "pytest>=7.0",
  "pytest-cov>=4.0",
  "ruff>=0.5.0",
  "black>=24.0",
  "mypy>=1.5",
  "pre-commit>=3.7.0",
  "httpx>=0.27.0"
]

[tool.pytest.ini_options]
addopts = "-q"
testpaths = ["tests"]
pythonpath = ["src"]

[tool.ruff]
line-length = 100
select = ["E", "F", "W", "I"]

[tool.ruff.format]
quote-style = "preserve"
indent-style = "space"

[tool.black]
line-length = 100

[tool.mypy]
python_version = "3.9"
strict = false
warn_unused_configs = true
warn_redundant_casts = true
warn_unused_ignores = true
ignore_missing_imports = true
